+++ src/components/input-forms/input-message-card/input-message-card.module.css
@@ -1,0 +1,6 @@
+.Form {
+    display: flex;
+    flex-direction: row-reverse;
+    justify-content: center;
+    align-items: center;
+}
